added SD_NETWORKD_CONFIG (#22)
$SD_NETWORKD_CONFIG (and its drop-ins) are copied to initramfs.

Intentionally, we do NOT fall back to a default if SD_NETWORKD_CONFIG is not
set.  That is because there is a tiny chance for a breaking change in case
someone uses a customized networkd.conf AND relies on the current behaviour
that this configuration file is NOT copied to the initramfs image. We try to
detect this situation and display a warning.

Later we will introduce a default for SD_NETWORKD_CONFIG
(/etc/systemd/networkd.conf) (and remove the warning again). This way
systemd-networkd configuration will always be copied to the initramfs image.